move external entity file handling to Document since it's used in Subset and DocType
